refactor!: Simplify folder structure for meta provider
