--- providers/openfeature-meta_provider/lib/open_feature/sdk/provider/version.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module OpenFeature
-  module SDK
-    module Provider
-      VERSION = "0.0.1"
-    end
-  end
-end
