+++ providers/openfeature-meta_provider/lib/openfeature/meta_provider_version.rb
@@ -1,0 +1,3 @@
+module OpenFeature
+  META_PROVIDER_VERSION = "0.0.1"
+end
